+++ src/convertSchema.ts
@@ -52,6 +52,16 @@ const flagsToSql = (flags?: string[]) =>
 const renderCommaIf = (cond: boolean) => (cond ? ',' : '')
 const padding = '  '
 
+const columnPolicy = (policy: string) => ` WITH (column_policy = '${policy}')`
+
+const getObjectPolicy = (node: Node, strictMode: boolean) => {
+  const noAdditionalProps = node.val.additionalProperties === false
+  if (strictMode && noAdditionalProps) {
+    return 'strict'
+  }
+  return 'dynamic'
+}
+
 const _convertSchema = (
   nodes: Node[],
   strictMode = false,
@@ -71,9 +81,7 @@ const _convertSchema = (
         'CREATE TABLE IF NOT EXISTS %s (\n' +
         _convertSchema(nodes.slice(1), strictMode, padding) +
         ')' +
-        (strictMode && node.val.additionalProperties
-          ? " WITH (column_policy = 'dynamic')"
-          : '')
+        columnPolicy(getObjectPolicy(node, strictMode))
       )
     }
     // Scalar fields, including objects with no defined fields
@@ -97,13 +105,7 @@ const _convertSchema = (
       const sqlType =
         node.val.bsonType === 'array'
           ? 'ARRAY'
-          : `OBJECT(${
-              strictMode
-                ? node.val.additionalProperties === false
-                  ? 'STRICT'
-                  : 'DYNAMIC'
-                : 'IGNORED'
-            }) AS`
+          : `OBJECT(${getObjectPolicy(node, strictMode).toUpperCase()}) AS`
       returnVal +=
         `${spacing}${field}${sqlType} (\n` +
         _convertSchema(childNodes, strictMode, newSpacing) +